WIP: refactor summary formatting - extract from index.js to dedicated module
- Move formatGateResults() logic from index.js to src/summary-adapter.js
- Index.js now calls renderCheckSummary() for thin orchestrator pattern
- Add formatRunSummaryJSON() for debugging (outputs raw JSON)
- All tests pass - functionality preserved

Next: can use formatRunSummaryJSON() to debug AI gate observations issue.
