+++ index.js
@@ -4,6 +4,7 @@
 import { loadRepoSpec } from './src/spec-loader.js';
 import { runAllGates } from './src/gates/index.js';
 import { postPRCommentWithGuards } from './src/pr-comment.js';
+import { renderCheckSummary } from './src/summary-adapter.js';
 
 const PR_REVIEW_NAME = "Cogni Git PR Review";
 
@@ -51,7 +52,7 @@ export default (app) => {
 
   async function createCompletedCheck(context, runResult, headSha, startTime) {
     const conclusion = mapStatusToConclusion(runResult.overall_status);
-    const { summary, text } = formatGateResults(runResult);
+    const { summary, text } = renderCheckSummary(runResult);
 
     const checkResult = await context.octokit.checks.create(context.repo({
       name: PR_REVIEW_NAME,
@@ -80,82 +81,6 @@ export default (app) => {
     }
   }
 
-  /**
-   * Format gate results into summary and text for GitHub check
-   */
-  function formatGateResults(runResult) {
-    const { gates, early_exit, duration_ms } = runResult;
-    
-    const failedGates = gates.filter(g => g.status === 'fail');
-    const neutralGates = gates.filter(g => g.status === 'neutral');
-    const passedGates = gates.filter(g => g.status === 'pass');
-    
-    // Summary
-    let summary;
-    if (gates.length === 0) {
-      summary = 'No gates configured';
-    } else if (failedGates.length > 0) {
-      summary = `Gate failures: ${failedGates.length}`;
-    } else if (neutralGates.length > 0) {
-      const reasons = [...new Set(neutralGates.map(g => g.neutral_reason).filter(Boolean))];
-      summary = `Gates neutral: ${reasons.join(', ')}`;
-    } else {
-      summary = 'All gates passed';
-    }
-    
-    if (early_exit) {
-      summary += ' (early exit)';
-    }
-    
-    // Text - detailed breakdown
-    let text = `Gates: ${gates.length} total | Duration: ${duration_ms}ms\n\n`;
-    
-    // Show gate status breakdown
-    text += `✅ Passed: ${passedGates.length} | ❌ Failed: ${failedGates.length} | ⚠️ Neutral: ${neutralGates.length}\n\n`;
-    
-    // Show failed gates first
-    if (failedGates.length > 0) {
-      text += '**Failures:**\n';
-      failedGates.forEach(gate => {
-        text += `• **${gate.id}**: ${gate.violations.length} violation(s)\n`;
-        gate.violations.slice(0, 3).forEach(v => { // Limit violations shown
-          text += `  - ${v.code}: ${v.message}\n`;
-        });
-        if (gate.violations.length > 3) {
-          text += `  - ...and ${gate.violations.length - 3} more\n`;
-        }
-      });
-      text += '\n';
-    }
-    
-    // Show neutral gates
-    if (neutralGates.length > 0) {
-      text += '**Neutral:**\n';
-      neutralGates.forEach(gate => {
-        text += `• **${gate.id}**: ${gate.neutral_reason || 'reason unknown'}\n`;
-      });
-      text += '\n';
-    }
-    
-    // Show passed gates summary
-    if (passedGates.length > 0) {
-      text += `**Passed:** ${passedGates.map(g => g.id).join(', ')}\n\n`;
-    }
-    
-    // Add stats from review limits if available
-    const reviewGate = gates.find(g => g.id === 'review_limits');
-    if (reviewGate?.stats) {
-      text += `**Stats:** files=${reviewGate.stats.changed_files || 0} | diff_kb=${reviewGate.stats.total_diff_kb || 0}`;
-    }
-    
-    // Add AI score if available. Temporary, only functional when only 1 AI rule. gate output needs refactoring. 
-    const rulesGate = gates.find(g => g.id === 'rules');
-    if (rulesGate?.stats?.score !== undefined) {
-      text += reviewGate?.stats ? ` | AI score=${rulesGate.stats.score}` : `**Stats:** AI score=${rulesGate.stats.score}`;
-    }
-    
-    return { summary, text };
-  }
 
   async function handlePullRequest(context) {
     const pr = context.payload.pull_request;
